自定义TextView

+++ 新浪微博/Main【主要】/Home【首页】/Model/Weico.m
@@ -74,9 +74,9 @@
             return [fmt stringFromDate:createDate];
         } else if ([createDate isToday]) { // 今天
             if (cmps.hour >= 1) {
-                return [NSString stringWithFormat:@"%d小时前", cmps.hour];
+                return [NSString stringWithFormat:@"%d小时前", (int)cmps.hour];
             } else if (cmps.minute >= 1) {
-                return [NSString stringWithFormat:@"%d分钟前", cmps.minute];
+                return [NSString stringWithFormat:@"%d分钟前", (int)cmps.minute];
             } else {
                 return @"刚刚";
             }
