+++ 新浪微博/Main【主要】/Home【首页】/View/WeicoPhotosView.m
@@ -107,7 +107,7 @@
     
     // 2.设置图片浏览器显示的所有图片
     NSMutableArray *photos = [NSMutableArray array];
-    int count = self.photos.count;
+    NSUInteger count = self.photos.count;
     for (int i = 0; i<count; i++) {
         Photo *pic = self.photos[i];
